requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
structlog>=24.1.0
pytest>=8.0.0
httpx>=0.27.0
python-dotenv>=1.0.1
pytest-asyncio>=0.23.5
pymupdf>=1.23.0
python-multipart>=0.0.9
pydicom>=2.4.4
Pillow>=10.3.0
pytesseract>=0.3.13
openpyxl>=3.1.2
pandas>=2.2.1
filetype>=1.2.0
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35
spacy>=3.7.4
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0
scikit-learn>=1.4.1.post1
cerebras-cloud-sdk>=1.0.0
nibabel>=5.2.1